+++ shared-helpers/index.ts
@@ -24,6 +24,7 @@ export * from "./src/views/multiselectQuestions"
 export * from "./src/views/occupancyFormatting"
 export * from "./src/views/summaryTables"
 export * from "./src/views/forgot-password/FormForgotPassword"
+export * from "./src/views/layout/AlertBanner"
 export * from "./src/views/layout/ExygyFooter"
 export * from "./src/views/sign-in/FormSignIn"
 export * from "./src/views/sign-in/FormSignInPwdless"
