+++ shared-helpers/src/locales/es.json
@@ -41,6 +41,7 @@
   "account.signUpSaveTime.useACode": "Utilice un código para iniciar sesión sin contraseña",
   "account.viewApplications": "Ver solicitudes",
   "alert.applicationMessage": "¡Las solicitudes ya están disponibles para más listados! Es posible que deba crear una cuenta. Revise los <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Términos de uso</a> y la <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Política de privacidad</a>, los cuales han cambiado.",
+  "alert.experiencingIssues": "Actualmente tenemos un problema con la creación de cuentas y los correos electrónicos de confirmación de inicio de sesión. Mientras solucionamos este problema, aún puede realizar búsquedas de viviendas asequibles sin iniciar sesión en su cuenta.\n\nRevise los <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Términos de uso</a> y la <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Política de privacidad</a>, los cuales han cambiado.",
   "alert.maintenance": "Este sitio está en mantenimiento programado. Nos disculpamos por cualquier inconveniente.",
   "application.ada.hearing": "Para deficiencias auditivas",
   "application.ada.label": "Viviendas accesibles de conformidad con ADA",
