+++ shared-helpers/src/locales/tl.json
@@ -41,6 +41,7 @@
   "account.signUpSaveTime.useACode": "Gumamit ng code upang mag-sign in nang walang password",
   "account.viewApplications": "Tingnan ang mga aplikasyon",
   "alert.applicationMessage": "Ang mga aplikasyon ay magagamit na ngayon para sa higit pang mga listahan! Maaaring kailanganin kang gumawa ng account. Pakisuri ang <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Mga Tuntunin ng Paggamit</a> at <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Patakaran sa Privacy</a>, na parehong nagbago.",
+  "alert.experiencingIssues": "Kasalukuyan kaming nagkakaroon ng problema sa paggawa ng account at mga email sa pagkumpirma sa pag-log in. Habang niresolba namin ang isyung ito, maaari ka pa ring magsagawa ng mga paghahanap sa abot-kayang pabahay nang hindi nagsa-sign in sa iyong account.\n\nPakisuri ang <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Mga Tuntunin ng Paggamit</a> at <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Patakaran sa Privacy</a>, na parehong nagbago.",
   "alert.maintenance": "Ang site na ito ay sumasailalim sa naka-iskedyul na pagpapanatili. Humihingi kami ng paumanhin para sa anumang abala.",
   "application.ada.hearing": "Para sa mga Kapansanan sa Pandinig",
   "application.ada.label": "Mga Accessible Unit ng ADA",
