+++ shared-helpers/src/locales/vi.json
@@ -41,6 +41,7 @@
   "account.signUpSaveTime.useACode": "Sử dụng mã để đăng nhập mà không cần mật khẩu",
   "account.viewApplications": "Xem hồ sơ",
   "alert.applicationMessage": "Các ứng dụng hiện có sẵn cho nhiều danh sách hơn! Bạn có thể được yêu cầu tạo một tài khoản. Vui lòng xem lại <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Điều khoản sử dụng</a> và <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Chính sách quyền riêng tư</a>, cả hai đều đã thay đổi.",
+  "alert.experiencingIssues": "Chúng tôi hiện đang gặp sự cố với email xác nhận đăng nhập và tạo tài khoản. Trong khi chúng tôi giải quyết vấn đề này, bạn vẫn có thể tiến hành tìm kiếm nhà ở giá phải chăng mà không cần đăng nhập vào tài khoản của mình.\n\nVui lòng xem lại <a href='https://mtc.ca.gov/doorway-housing-portal-terms-use' target='_blank'>Điều khoản sử dụng</a> và <a href='https://mtc.ca.gov/doorway-housing-portal-privacy-policy' target='_blank'>Chính sách quyền riêng tư</a>, cả hai đều đã thay đổi.",
   "alert.maintenance": "Trang web này đang được bảo trì theo lịch trình. Chúng tôi xin lỗi vì bất cứ sự bất tiện nào.",
   "application.ada.hearing": "Đối với người khiếm thính",
   "application.ada.label": "Các Căn nhà Dễ tiếp cập của ADA",
