+++ sites/partners/.env.template
@@ -10,3 +10,7 @@ CLOUDINARY_KEY='abcxyz'
 CLOUDINARY_SIGNED_PRESET='test123'
 MAPBOX_TOKEN=
 SHOW_SMS_MFA=TRUE
+
+# YYYY-MM-DD HH:mm Z,YYYY-MM-DD HH:mm Z
+# 2024-02-05 18:00 -08:00,2024-02-05 18:55 -08:00
+MAINTENANCE_WINDOW=
