+++ sites/partners/src/layouts/index.tsx
@@ -1,4 +1,5 @@
 import React, { useContext } from "react"
+import Markdown from "markdown-to-jsx"
 import Head from "next/head"
 import { useRouter } from "next/router"
 import {
@@ -10,7 +11,7 @@ import {
   MenuLink,
   setSiteAlertMessage,
 } from "@bloom-housing/ui-components"
-import { AuthContext, ExygyFooter } from "@bloom-housing/shared-helpers"
+import { AlertBanner, AuthContext, ExygyFooter } from "@bloom-housing/shared-helpers"
 
 const Layout = (props) => {
   const { profile, signOut } = useContext(AuthContext)
@@ -54,6 +55,9 @@ const Layout = (props) => {
         <Head>
           <title>{t("nav.siteTitlePartners")}</title>
         </Head>
+        <AlertBanner maintenanceWindow={process.env.maintenanceWindow} variant={"alert"}>
+          <Markdown>{t("alert.experiencingIssues")}</Markdown>
+        </AlertBanner>
         <SiteHeader
           imageOnly={true}
           logoSrc="/images/doorway-logo-partners.svg"
