--- sites/public/src/layouts/application.module.scss
@@ -1,14 +1,0 @@
-.site-alert-banner-container {
-  // temp launch change, replace with --seeds-color-alert-light after advertising window
-  background-color: var(--seeds-color-primary-lighter);
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.site-alert-banner-content {
-  --common-message-max-width: var(--seeds-width-5xl);
-  --common-message-font-size: var(--seeds-font-size-base);
-  width: 100%;
-}
