+++ sites/public/src/layouts/application.tsx
@@ -1,14 +1,11 @@
 import React, { useContext } from "react"
-import dayjs from "dayjs"
+import Markdown from "markdown-to-jsx"
 import { useRouter } from "next/router"
 import Head from "next/head"
 import { MenuLink, setSiteAlertMessage, t } from "@bloom-housing/ui-components"
-import { AuthContext } from "@bloom-housing/shared-helpers"
+import { AlertBanner, AuthContext } from "@bloom-housing/shared-helpers"
 import { getSiteFooter } from "../lib/helpers"
 import { SiteHeader } from "@bloom-housing/doorway-ui-components/src/headers/SiteHeader"
-import { Message } from "@bloom-housing/ui-seeds"
-import styles from "./application.module.scss"
-import Markdown from "markdown-to-jsx"
 
 const Layout = (props) => {
   const { profile, signOut } = useContext(AuthContext)
@@ -108,19 +105,6 @@ const Layout = (props) => {
       })
     }
   }
-  const getInMaintenance = () => {
-    let inMaintenance = false
-    const maintenanceWindow = process.env.maintenanceWindow?.split(",")
-    if (maintenanceWindow?.length === 2) {
-      const convertWindowToDate = (windowString: string) =>
-        dayjs(windowString, "YYYY-MM-DD HH:mm Z")
-      const startWindow = convertWindowToDate(maintenanceWindow[0])
-      const endWindow = convertWindowToDate(maintenanceWindow[1])
-      const now = dayjs()
-      inMaintenance = now > startWindow && now < endWindow
-    }
-    return inMaintenance
-  }
 
   return (
     <div className="site-wrapper">
@@ -128,14 +112,10 @@ const Layout = (props) => {
         <Head>
           <title>{t("nav.siteTitle")}</title>
         </Head>
-        {getInMaintenance() && (
-          <div className={styles["site-alert-banner-container"]}>
-            <Message className={styles["site-alert-banner-content"]} variant={"primary"}>
-              {/* temporary change to support doorway launch, should return to alert.maintenance and variant=alert after advertisig window */}
-              <Markdown>{t("alert.applicationMessage")}</Markdown>
-            </Message>
-          </div>
-        )}
+        {/* temporary change to support email issue, should return to alert.maintenance after window */}
+        <AlertBanner maintenanceWindow={process.env.maintenanceWindow} variant={"alert"}>
+          <Markdown>{t("alert.experiencingIssues")}</Markdown>
+        </AlertBanner>
         <SiteHeader
           logoSrc="/images/doorway-logo.png"
           homeURL="/"
